Desktop/listed/requirements.txt
flask
python-dotenv
azure-cognitiveservices-vision-contentmoderator
mrest